requirements.txt
matplotlib==3.6.3
numpy==2.2.1
Pillow==9.0.1
Pillow==11.1.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
